requirements.txt
asgiref==3.7.2
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.40.0
Django==5.0.2
django-cloudinary-storage==0.3.0
django-js-asset==2.2.0
django-social-share==2.3.0
django-summernote==0.8.20.0
idna==3.6
mysqlclient==2.2.4
pillow==10.3.0
python-decouple==3.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==8.11.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
whitenoise==6.6.0
